pyproject.toml
[project]
name = "prism"
version = "0.1.0"
description = "AI-powered personal recommendation engine using pairwise comparison"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
keywords = ["ai", "recommendation", "pairwise", "preference-learning", "signal-intelligence"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: Chinese (Simplified)",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "fastapi>=0.100",
    "httpx>=0.25",
    "jinja2>=3.1",
    "python-dotenv>=1.0",
    "uvicorn>=0.30",
    "pyyaml>=6.0",
    "ruamel.yaml>=0.18",
    "click>=8.0",
    "markdown>=3.10.2",
    "python-multipart>=0.0.26",
    "ebooklib>=0.18",
    "omlx-sdk",
    "youtube-transcript-api>=1.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]
thread = ["playwright>=1.40"]
ctr = [
    "xgboost>=2.0",
    "pandas>=2.0",
    "numpy>=1.26",
    "scikit-learn>=1.3",
]

[project.urls]
Homepage = "https://github.com/xiaohongsimon/prism"
Repository = "https://github.com/xiaohongsimon/prism"
Issues = "https://github.com/xiaohongsimon/prism/issues"

[project.scripts]
prism = "prism.cli:cli"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

# Only package the `prism` source tree. The repo also contains a sibling
# `prism-lens/` and top-level `tests/` that must not be auto-discovered.
[tool.setuptools.packages.find]
include = ["prism*"]
exclude = ["prism-lens*", "tests*", "scripts*", "data*", "design*", "docs*"]

# Install the local omlx-sdk checkout as an editable dependency so changes
# to the sibling project reflect immediately without re-syncing.
[tool.uv.sources]
omlx-sdk = { path = "../omlx-manager/sdk", editable = true }
